RUMM-180 Extract all utils to core
