+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/file/FileOrchestrator.kt
@@ -7,7 +7,7 @@
 package com.datadog.android.core.internal.data.file
 
 import com.datadog.android.core.internal.data.Orchestrator
-import com.datadog.android.log.internal.utils.sdkLogger
+import com.datadog.android.core.internal.utils.sdkLogger
 import java.io.File
 import java.io.FileFilter
