+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/file/ImmediateFileWriter.kt
@@ -9,8 +9,8 @@ package com.datadog.android.core.internal.data.file
 import com.datadog.android.core.internal.data.Orchestrator
 import com.datadog.android.core.internal.data.Writer
 import com.datadog.android.core.internal.domain.Serializer
-import com.datadog.android.log.internal.utils.devLogger
-import com.datadog.android.log.internal.utils.sdkLogger
+import com.datadog.android.core.internal.utils.devLogger
+import com.datadog.android.core.internal.utils.sdkLogger
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.FileOutputStream
