+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/utils/WorkManagerUtils.kt
@@ -7,7 +7,6 @@ import androidx.work.NetworkType
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkManager
 import com.datadog.android.core.internal.data.upload.UploadWorker
-import com.datadog.android.log.internal.utils.sdkLogger
 import java.lang.IllegalStateException
 
 internal const val TAG = "WorkManagerUtils"
