+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/utils/RuntimeUtilsTest.kt
@@ -1,4 +1,4 @@
-package com.datadog.android.log.internal.utils
+package com.datadog.android.core.internal.utils
 
 import android.util.Log
 import com.datadog.android.BuildConfig
